+++ src/workbench/tree-data-providers/tree-items/graphos-supergraphs/notLoggedInTreeItem.ts
@@ -1,6 +1,6 @@
 import { TreeItem, TreeItemCollapsibleState } from 'vscode';
 
-export class NotLoggedInTreeItem extends TreeItem {
+export default class NotLoggedInTreeItem extends TreeItem {
   constructor() {
     super('Login with GraphOS', TreeItemCollapsibleState.None);
     this.command = {